[refactoring] 라이딩 경로 컴포넌트 분리
