+++ src/components/Users/RidingRoute/items/Main/RouteInfo.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Link } from "react-router-dom";
 // TODO route path 요청
-// import { getRidingRouteById } from "../../../../api/RidingRoute";
-import { getTimeContext } from "../../../../util/getDateContext";
+// import { getRidingRouteById } from "../../../../../api/RidingRoute";
+import { getTimeContext } from "../../../../../util/getDateContext";
 
 import "./RouteInfo.css";
